Do the 'if err != nil' checks in E rather than e
to allow partial inlining.

Call runtime.CallersFrames lazily.

    name       old time/op    new time/op    delta
    Success-8    4.40ns ± 1%    3.79ns ± 1%  -13.83%  (p=0.000 n=10+10)
    Failure-8     477ns ± 0%     312ns ± 1%  -34.68%  (p=0.000 n=9+10)

    name       old alloc/op   new alloc/op   delta
    Success-8     0.00B          0.00B          ~     (all equal)
    Failure-8      312B ± 0%       24B ± 0%  -92.31%  (p=0.000 n=10+10)

    name       old allocs/op  new allocs/op  delta
    Success-8      0.00           0.00          ~     (all equal)
    Failure-8      3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)

Fixes #11

Co-authored-by: Jeff Wendling <jeff@storj.io>